Use production-grade webserver

+++ hydra-tui/src/Hydra/Client.hs
@@ -62,7 +62,7 @@ withClient Options{hydraNodeHost = Host{hostname, port}, cardanoSigningKey} call
         }
  where
   -- TODO(SN): ping thread?
-  client q = runClient (toString hostname) (fromIntegral port) "/" $ \con -> do
+  client q = runClient (toString hostname) port "/" $ \con -> do
     -- REVIEW(SN): is sharing the 'con' fine?
     callback ClientConnected
     race_ (receiveOutputs con) (sendInputs q con)
